backend/requirements.txt
blinker==1.9.0
gunicorn
boto3==1.34.0
botocore==1.34.162
click==8.3.1
colorama==0.4.6
Flask==3.0.0
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
MarkupSafe==3.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
s3transfer==0.9.0
six==1.17.0
urllib3==2.5.0
Werkzeug==3.1.4
